+++ package.json
@@ -53,7 +53,8 @@
     "@plone/registry": "workspace:*",
     "@plone/types": "workspace:*",
     "@plone/components": "workspace:^",
-    "@kitconcept/volto-logos-block": "workspace:*"
+    "@kitconcept/volto-logos-block": "workspace:*",
+    "@kitconcept/volto-light-theme": "workspace:*"
   },
   "devDependencies": {
     "mrs-developer": "^2.2.0"
