--- packages/volto-logos-block/src/components/index.ts
@@ -1,4 +1,0 @@
-import LogoBlockEdit from './Logo/Edit';
-import LogoBlockView from './Logo/View';
-
-export { LogoBlockEdit, LogoBlockView };
