--- packages/volto-logos-block/src/components/widgets/index.ts
@@ -1,5 +1,0 @@
-import BlocksObjectWidget from './BlocksObjectWidget';
-import LogoBlockSizeWidget from './LogoBlockSizeWidget';
-import LogoBlockWidthWidget from './LogoBlockWidthWidget';
-
-export { BlocksObjectWidget, LogoBlockSizeWidget, LogoBlockWidthWidget };
