+++ volto-logos-block/babel.config.js
