+++ volto-logos-block/locales/volto.pot
