+++ volto-logos-block/package.json
