+++ volto-logos-block/src/Theme/_main.scss
@@ -12,3 +12,9 @@
 .content-area .block.logos a.external::after {
   content: none;
 }
+
+// .q.container.layout, .a.container.layout, .ui.container.layout {
+//   max-width: var(--layout-container-width);
+//   margin-right: auto;
+//   margin-left: auto;
+// }
