+++ volto-logos-block/src/components/View.tsx
@@ -3,13 +3,12 @@ import { Message } from 'semantic-ui-react';
 import { defineMessages, useIntl } from 'react-intl';
 import cx from 'classnames';
 import isEmpty from 'lodash/isEmpty';
-
+import type { BlockViewProps } from '@plone/types';
 import { flattenToAppURL } from '@plone/volto/helpers';
 import { Container } from '@plone/components';
 import imageBlockSVG from '@plone/volto/components/manage/Blocks/Image/block-image.svg';
 import UniversalLink from '@plone/volto/components/manage/UniversalLink/UniversalLink';
-import type { logoData } from '@kitconcept/volto-logos-block/types/Logo';
-
+import type { logoData } from '@kitconcept/volto-logos-block/components/Edit';
 const messages = defineMessages({
   PleaseChooseLogo: {
     id: 'Please choose a logo as source for this element',
@@ -17,16 +16,13 @@ const messages = defineMessages({
   },
 });
 
-interface logosViewProps {
+interface logosViewProps extends BlockViewProps {
   '@type': string;
   data: logoData;
   isEditMode: boolean;
 }
 
-const View: React.FC<logosViewProps> = (props: {
-  data: logoData;
-  isEditMode: boolean;
-}) => {
+const View = (props: logosViewProps) => {
   const intl = useIntl();
   const { isEditMode } = props;
