+++ volto-logos-block/src/components/index.ts
@@ -1,0 +1,4 @@
+import LogoBlockEdit from './Edit';
+import LogoBlockView from './View';
+
+export { LogoBlockEdit, LogoBlockView };
